Suppress warning in 2d-regional-scikit

+++ src/methods.py
@@ -183,7 +183,7 @@ def scikit2dr(args, colourimg, greyimg):
     # From scikit docs:
     # The entropy is computed using base 2 logarithm i.e. the filter returns
     # the minimum number of bits needed to encode the local gray level distribution.
-    entimg = skentropy(greyimg, skdisk(args.radius))
+    entimg = skentropy(greyimg.astype(np.uint8), skdisk(args.radius))
     entropy = entimg.mean()
 
     log(f"entropy: {entropy}")
